requirements.txt
opencv_python==4.5.1.48
imutils==0.5.4
fuzzywuzzy==0.18.0
numpy==1.19.5
hickle==4.0.4
mysql-connector-python
Pillow==8.1.0



